Various changes, noted below.
Attempted to introduce a "predist" targe in the Makefile in order to
deal with prereleases of a version. Mixed results.

Modified the various src/fqu_*/Makefile.am to change relative paths
in the -I options into using $(top_srcdir). This allows builds to
occur in a separate location, and now "make distcheck" completes
successfully.